% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.spv.R, R/print.spvforlist.R,
%   R/print.spvlist.R, R/print.spvlistforlist.R
\name{print.spv}
\alias{print.spv}
\alias{print.spvlist}
\alias{print.spvforlist}
\alias{print.spvlistforlist}
\title{Print Method for S3 \code{spv} classes}
\usage{
\method{print}{spv}(x, ...)

\method{print}{spvforlist}(x, ...)

\method{print}{spvlist}(x, ...)

\method{print}{spvlistforlist}(x, ...)
}
\arguments{
\item{x}{Object of class \code{spv} or \code{spvlist}}

\item{\dots}{Unimplemented}
}
\description{
Simple print methods for S3 classes \code{spv}, \code{spvlist}, \code{spvforlist} and \code{spvlistforlist}. See 
\code{\link{plot.spv}} for examples.
}
\references{
Pieter C. Schoonees, Niel J. le Roux, Roelof L.J. Coetzer (2016). Flexible Graphical Assessment of 
Experimental Designs in R: The vdg Package. \emph{Journal of Statistical Software}, 74(3), 1-22. 
\doi{10.18637/jss.v074.i03}.
}
\author{
Pieter C. Schoonees
}
\keyword{print}
