% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCDH5.R
\docType{data}
\name{SCDH5}
\alias{SCDH5}
\title{Design from the Vdgraph package}
\format{
a data frame of 28 runs in four variables.
}
\usage{
SCDH5
}
\description{
This data frame is taken verbatim from the (now archived) \pkg{Vdgraph} package. See
that package for the original reference.
}
\references{
Lawson J, Vining G (2014). \emph{Vdgraph: Variance dispersion graphs and Fraction of design space plots for response surface designs}. R package version 2.2-2, \url{https://CRAN.R-project.org/package=Vdgraph}.
}
