% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vdg-package.R
\docType{package}
\name{vdg-package}
\alias{vdg}
\alias{vdg-package}
\title{Variance Dispersion Graphs (VDGs) and Fraction of Design Space (FDS) Plots}
\description{
Produces VDG, FDS and related plots for assessing experimental designs.
}
\author{
Pieter C. Schoonees <schoonees@gmail.com>
}
\keyword{package}

